Add `tests/units/utils.rs` to `platform_pki_2_electric_boogaloo`
- Introduce `initialize_pki_system` to create the `PkiSystem` object and ensure
  test certificates are available in the OS certificate store
- `InstalledCertificates` fixture provides sync methods to obtain the certificate reference
  without doing extra santity check (already done done `PkiSystem` initialization)
- Remove `InstalledCertificates::alice_sign_message` method: no longer used and complex
  to keep since it requires platform-specific signature code.
  Instead the tests should either use a pre-generated signature, or initialize their
  own `PkiSystem` to build a signature.
- Add `InstalledCertificates::alice_encrypt_message` method: this helper is possible
  since encryption uses shared encryption code so it requires no `PkiSystem` initialization.
